LFS-22 / LFS-115: Cleanup the serverside codebase [misc] Split the codebase into spi and internal packages
